python-backend/requirements.txt
openai-agents
openai-chatkit
pydantic
fastapi
uvicorn
python-dotenv

# Local dspy-guardrails (monorepo)
-e ../../dspyGuardrails
